chore: reuse alloy genesis in bindings
