+++ crates/node-bindings/Cargo.toml
@@ -13,9 +13,9 @@ exclude.workspace = true
 
 [dependencies]
 alloy-primitives = { workspace = true, features = ["k256", "serde"] }
+alloy-genesis.workspace = true
 k256.workspace = true
 serde_json.workspace = true
-serde.workspace = true
 tempfile.workspace = true
 
 [dev-dependencies]
